Fix and simplify Python example
