--- examples/python/auto-tracing/.gitignore
@@ -1,3 +1,0 @@
-.logfire
-logfire_credentials.json
-venv
