--- examples/python/auto-tracing/app/main.py
@@ -1,11 +1,0 @@
-from app.server import create_app
-
-
-def main():
-    """Entry point into app."""
-    app = create_app()
-    app.run(debug=True)
-
-
-if __name__ == '__main__':
-    main()
