+++ examples/python/auto-tracing/app/server/templates/history.html
@@ -28,6 +28,6 @@
             {% endfor %}
         </tbody>
     </table>
-    <a href="{{ url_for('main.index') }}">Back to Calculator</a>
+    <a href="{{ url_for('index') }}">Back to Calculator</a>
 </body>
 </html>
