+++ examples/python/auto-tracing/app/server/templates/index.html
@@ -17,7 +17,7 @@
         <button onclick="calculate()">Calculate</button>
     </div>
     <div id="result"></div>
-    <a href="{{ url_for('main.history') }}">View Calculation History</a>
+    <a href="{{ url_for('history') }}">View Calculation History</a>
 
     <script>
         const inputFields = {
